+++ packages/react-components/react-nav-preview/src/components/NavSubItem/index.ts
@@ -1,0 +1,5 @@
+export { NavSubItem } from './NavSubItem';
+export type { NavSubItemSlots, NavSubItemProps, NavSubItemState } from './NavSubItem.types';
+export { renderNavSubItem_unstable } from './renderNavSubItem';
+export { useNavSubItem_unstable } from './useNavSubItem';
+export { useNavSubItemStyles_unstable, navSubItemClassNames } from './useNavSubItemStyles.styles';