+++ packages/react-components/react-nav-preview/src/components/useNavContextValues.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { NavContextValue, NavContextValues, NavState } from '../Nav';
 
 export function useNavContextValues_unstable(state: NavState): NavContextValues {
-  const { selectedValue, onRegister, onUnregister, onSelect, getRegisteredNavGroups } = state;
+  const { selectedValue, onRegister, onUnregister, onSelect, getRegisteredNavItems } = state;
 
   const navContext = React.useMemo<NavContextValue>(
     () => ({
@@ -10,9 +10,9 @@ export function useNavContextValues_unstable(state: NavState): NavContextValues 
       onSelect,
       onRegister,
       onUnregister,
-      getRegisteredNavGroups,
+      getRegisteredNavItems,
     }),
-    [selectedValue, onSelect, onRegister, onUnregister, getRegisteredNavGroups],
+    [selectedValue, onSelect, onRegister, onUnregister, getRegisteredNavItems],
   );
 
   return { nav: navContext };