+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverButton/TeachingPopoverButton.tsx
@@ -1,20 +1,23 @@
 import * as React from 'react';
-import type { ForwardRefComponent } from '@fluentui/react-utilities';
-import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
-import { useTeachingPopoverButton_unstable } from './useTeachingPopoverButton';
 import { renderTeachingPopoverButton_unstable } from './renderTeachingPopoverButton';
+import { useTeachingPopoverButton_unstable } from './useTeachingPopoverButton';
 import { useTeachingPopoverButtonStyles_unstable } from './useTeachingPopoverButtonStyles.styles';
+import type { ForwardRefComponent } from '@fluentui/react-utilities';
 import type { TeachingPopoverButtonProps } from './TeachingPopoverButton.types';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
 
 /**
- * TeachingPopoverButton component - TODO: add more docs
+ * TeachingPopoverButtons override style and functionality based on carousel/popover state.
  */
 export const TeachingPopoverButton: ForwardRefComponent<TeachingPopoverButtonProps> = React.forwardRef((props, ref) => {
   const state = useTeachingPopoverButton_unstable(props, ref);
 
   useTeachingPopoverButtonStyles_unstable(state);
+
   useCustomStyleHook_unstable('useTeachingPopoverButtonStyles_unstable')(state);
+
   return renderTeachingPopoverButton_unstable(state);
-});
+  // Casting is required due to lack of distributive union to support unions on @types/react
+}) as ForwardRefComponent<TeachingPopoverButtonProps>;
 
 TeachingPopoverButton.displayName = 'TeachingPopoverButton';