+++ packages/react-components/react-teaching-popover-preview/stories/TeachingPopoverActions/TeachingPopoverActionsDefault.stories.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { TeachingPopoverActions, TeachingPopoverActionsProps } from '@fluentui/react-teaching-popover-preview';
-
-export const Default = (props: Partial<TeachingPopoverActionsProps>) => <TeachingPopoverActions {...props} />;