+++ packages/react-components/react-teaching-popover-preview/stories/TeachingPopoverSurface/TeachingPopoverSurfaceDefault.stories.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { TeachingPopoverSurface, TeachingPopoverSurfaceProps } from '@fluentui/react-teaching-popover-preview';
-
-export const Default = (props: Partial<TeachingPopoverSurfaceProps>) => <TeachingPopoverSurface {...props} />;