+++ packages/azure-themes/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/azure-themes
 
-This log was last generated on Thu, 14 Dec 2023 07:30:09 GMT and should not be manually modified.
+This log was last generated on Mon, 25 Dec 2023 07:28:36 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.6.49](https://github.com/microsoft/fluentui/tree/@fluentui/azure-themes_v8.6.49)
+
+Mon, 25 Dec 2023 07:28:36 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/azure-themes_v8.6.48..@fluentui/azure-themes_v8.6.49)
+
+### Patches
+
+- Bump @fluentui/react to v8.114.0 ([PR #30158](https://github.com/microsoft/fluentui/pull/30158) by beachball)
 
 ## [8.6.48](https://github.com/microsoft/fluentui/tree/@fluentui/azure-themes_v8.6.48)